requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py38h32f6830_1001
argon2-cffi=20.1.0=py38h4d0b108_1
async_generator=1.10=py_0
attrs=20.2.0=pyh9f0ad1d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.2.1=pyh9f0ad1d_0
ca-certificates=2020.7.22=0
certifi=2020.6.20=py38_0
cffi=1.14.3=py38hed5b41f_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
coverage=5.3=py38haf1e3a3_0
cycler=0.10.0=pypi_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py38h32f6830_1001
filelock=3.0.12=pypi_0
freetype=2.10.2=ha233b18_0
gdown=3.12.2=pypi_0
idna=2.10=pypi_0
importlib-metadata=2.0.0=py38h32f6830_0
importlib_metadata=2.0.0=0
iniconfig=1.0.1=pypi_0
intel-openmp=2019.4=233
ipykernel=5.3.4=py38h23f93f0_0
ipython=7.18.1=py38h1cdfbd6_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.5.1=pypi_0
jedi=0.15.2=py38_0
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.16.0=pypi_0
jpeg=9b=he5867d9_2
jsonschema=3.2.0=py38h32f6830_1
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py38h32f6830_1
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
kiwisolver=1.2.0=pypi_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=haf1e3a3_1
libtiff=4.1.0=hcb84e12_1
lz4-c=1.9.2=hb1e8313_1
markupsafe=1.1.1=py38h4d0b108_1
matplotlib=3.3.2=pypi_0
mistune=0.8.4=py38h64e0658_1001
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
nbclient=0.5.0=py_0
nbconvert=6.0.6=py38h32f6830_0
nbformat=5.0.7=py_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.4.1=py_0
ninja=1.10.1=py38h879752b_0
notebook=6.1.4=py38h32f6830_0
numpy=1.19.2=pypi_0
numpy-base=1.19.1=py38hcfb5961_0
olefile=0.46=py_0
openssl=1.1.1h=haf1e3a3_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.2=pypi_0
pandoc=2.10.1=haf1e3a3_0
pandocfilters=1.4.2=py_1
parso=0.5.2=py_0
pexpect=4.8.0=py38h32f6830_1
pickleshare=0.7.5=py38h32f6830_1001
pillow=7.2.0=py38ha54b6ba_0
pip=20.2.2=py38_0
pluggy=0.13.1=pypi_0
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.7=py_0
ptyprocess=0.6.0=py_1001
py=1.9.0=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py38h4d0b108_0
pysocks=1.7.1=pypi_0
pytest=6.1.1=pypi_0
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_0
pytz=2020.1=pypi_0
pyzmq=19.0.2=py38hb1e8313_1
readline=8.0=h1de35cc_0
regex=2020.9.27=pypi_0
requests=2.24.0=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.2=pypi_0
send2trash=1.5.0=py_0
sentencepiece=0.1.91=pypi_0
setuptools=49.6.0=py38_1
six=1.15.0=pyh9f0ad1d_0
sklearn=0.0=pypi_0
sqlite=3.33.0=hffcf06c_0
terminado=0.9.1=py38h32f6830_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hb0a8c7a_0
tokenizers=0.8.1rc2=pypi_0
toml=0.10.1=pypi_0
torchvision=0.7.0=py38_cpu
tornado=6.0.4=py38h64e0658_1
tqdm=4.50.0=pypi_0
traitlets=5.0.4=py_1
transformers=3.3.1=pypi_0
urllib3=1.25.10=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wget=3.2=pypi_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.2=hb1e8313_3
zipp=3.2.0=py_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
